refactor: RequestBodyArgumentResolver 상속 제거
AbstractModelArgumentResolver를 상속받아 확장하던 구조를
애노테이션 기반 ArgumentResolver임을 명시하기 위해
AbstractAnnotationArgumentResolver를 확장하도록 변경함
